// frontend/src/types/bookingRequestStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type BookingRequestStatus = typeof BookingRequestStatus[keyof typeof BookingRequestStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BookingRequestStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED',
  COMPLETED: 'COMPLETED',
} as const;