// frontend/src/types/bookingResponse.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { BookingResponseType } from './bookingResponseType';
import type { BookingResponseStatus } from './bookingResponseStatus';

export interface BookingResponse {
  id?: number;
  version?: number;
  createdBy?: number;
  updatedBy?: number;
  createdAt?: string;
  updatedAt?: string;
  name?: string;
  email?: string;
  phone?: string;
  type?: BookingResponseType;
  note?: string;
  date?: string;
  time?: string;
  status?: BookingResponseStatus;
}